backend/pyproject.toml
[project]
name = "vigil-backend"
version = "2.0.0"
description = "Vigil backend server — plugin-based text analysis API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "litellm>=1.81.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.1.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[tool.uv]
dev-dependencies = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
